MdeModulePkg HiiDatabaseDxe: Fix build error on GCC 4.4
Remove extra left side (lvalue) type casting that generated
warnings with GCC 4.4.

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@11267 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/HiiDatabaseDxe/String.c
@@ -801,9 +801,9 @@ InsertLackStringBlock (
   if (FrontSkipNum > 0) {
     *BlockPtr = *BlockType;
     if (*BlockType == EFI_HII_SIBT_SKIP1) {
-      (UINT8) (*(BlockPtr + sizeof (EFI_HII_STRING_BLOCK))) = (UINT8)FrontSkipNum;
+      *(BlockPtr + sizeof (EFI_HII_STRING_BLOCK)) = (UINT8) FrontSkipNum;
     } else {
-      (UINT16) (*(UINT16 *)(BlockPtr + sizeof (EFI_HII_STRING_BLOCK))) = (UINT16)FrontSkipNum;
+      *(UINT16 *)(BlockPtr + sizeof (EFI_HII_STRING_BLOCK)) = (UINT16) FrontSkipNum;
     }
     BlockPtr += SkipLen;
   }
@@ -822,9 +822,9 @@ InsertLackStringBlock (
   if (IdCount > FrontSkipNum + 1) {
     *BlockPtr = *BlockType;
     if (*BlockType == EFI_HII_SIBT_SKIP1) {
-      (UINT8) (*(BlockPtr + sizeof (EFI_HII_STRING_BLOCK))) = (UINT8) (IdCount - FrontSkipNum - 1);
+      *(BlockPtr + sizeof (EFI_HII_STRING_BLOCK)) = (UINT8) (IdCount - FrontSkipNum - 1);
     } else {
-      (UINT16) (*(UINT16 *)(BlockPtr + sizeof (EFI_HII_STRING_BLOCK))) = (UINT16) (IdCount - FrontSkipNum - 1);
+      *(UINT16 *)(BlockPtr + sizeof (EFI_HII_STRING_BLOCK)) = (UINT16) (IdCount - FrontSkipNum - 1);
     }
     BlockPtr += SkipLen;
   }
